scanner/requirements.txt
scapy
requests
python-dotenv
netaddr
